requirements.txt
django==2.2
mysqlclient
pillow
django-pure-pagination
redis
django-simple-captcha
django-crispy-forms
django-import-export
django-reversion
django-formtools
future
httplib2
six
xlwt
xlsxwriter
requests
twilio
django-simple-captcha
qrcode